requirements.txt
numpy==1.18.3
pandas==1.0.3
pytest==5.2.0
python-dateutil==2.8.1
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.19
sqlparse==0.3.0
urllib3==1.25.3